Extract dashboard cache scope to shared module
